+++ libs/rxjs/number/src/lib/parse-float.ts
@@ -20,7 +20,7 @@ import { map } from 'rxjs/operators';
  * ```
  *
  * @returns Float number from source string {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable}
- * @category RxJS Number Parsing
+ * @category RxJS Number Creation
  */
 export function parseFloat(): OperatorFunction<string, number> {
   return (source: Observable<string>) => source.pipe(map((value) => Number.parseFloat(value)));