+++ libs/rxjs/string/src/lib/filter-includes.spec.ts
@@ -1,12 +1,12 @@
 import { from } from 'rxjs';
-import { fromIncludes } from './from-includes';
+import { filterIncludes } from './filter-includes';
 import { reduce } from 'rxjs/operators';
 
 describe('fromIncludes', () => {
   it('should return a string if a string includes the value', (done) => {
     from(['test', 'testing', 'foobar'])
       .pipe(
-        fromIncludes('test'),
+        filterIncludes('test'),
         reduce((acc, val) => {
           acc.push(val);
           return acc;