+++ libs/rxjs/string/src/lib/filter-starts-with.spec.ts
@@ -1,12 +1,12 @@
 import { from } from 'rxjs';
 import { reduce, take } from 'rxjs/operators';
-import { fromStartsWith } from './from-starts-with';
+import { filterStartsWith } from './filter-starts-with';
 
 describe('fromStartsWith', () => {
   it('should return a string if a string ends with a character', (done) => {
     from(['test?', 'foo', 'test'])
       .pipe(
-        fromStartsWith('t'),
+        filterStartsWith('t'),
         reduce((acc, val) => {
           acc.push(val);
           return acc;