+++ libs/rxjs/string/src/lib/from-code-point.ts
@@ -2,32 +2,32 @@
  * @packageDocumentation
  * @module string
  */
-import { Observable, OperatorFunction } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Observable, SchedulerLike } from 'rxjs';
+import { scheduleCodePoint, subscribeToCodePoint } from '../utils/code-point';
 
 /**
- * The `fromCodePoint` operator can be used with an {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable} array of
- * numbers that represent character code points, and returns a string value
+ * The `fromCodePoint` operator is used to create an {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable} string
+ * from a number array of code points
  *
  * @remarks
- * Based on [String.fromCodePoint](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/fromCodePoint)
+ * Based on [String.fromCharCode](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/fromCharCode)
  * This operator takes an array of number values, if you want to only convert one character this still needs passed as an array
  *
  * @example
  * ```ts
- * from([9731, 9733, 9842])
- *  .pipe(
- *    reduce((acc, val) => {
- *      acc.push(val);
- *      return acc;
- *    }, []),
- *    fromCharCode()
- *  ).subscribe(...) // '☃★♲'
+ * fromCodePoint([9731, 9733, 9842]).subscribe(...) // '☃★♲'
  * ```
  *
  * @returns String from an array of character codes
- * @category RxJS From String Creation
+ * @category RxJS String Creation
  */
-export function fromCodePoint(): OperatorFunction<number[], string> {
-  return (source: Observable<number[]>) => source.pipe(map((values) => String.fromCodePoint(...values)));
+export function fromCodePoint(input: number[], scheduler?: SchedulerLike) {
+  if (!input) {
+    throw new Error('Input cannot be null');
+  }
+  if (!scheduler) {
+    return new Observable<string>(subscribeToCodePoint(input));
+  } else {
+    return scheduleCodePoint(input, scheduler);
+  }
 }